% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{div}
\alias{div}
\title{División de dos números}
\usage{
div(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número (distinto de 0)}
}
\value{
La división entre a y b
}
\description{
Función que realiza la división entre dos números
}
\examples{
mult(5,4)
}
\seealso{
\link{mult}
}
