% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Resta de dos números}
\usage{
resta(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número}
}
\value{
La resta entre a y b
}
\description{
Función que realiza la resta entre dos números
}
\examples{
resta(5,4)
}
\seealso{
\link{suma}
}
